requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-socketio[asyncio]==5.8.0
python-engineio==4.3.4 
cassandra-driver==3.29.2
aiohttp==3.11.12
PyJWT==2.10.1
pydantic-settings>=2.9.1
redis==4.3.4
